requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.18.0
numpy==1.26.4
pillow==10.2.0
opencv-python==4.8.1.78